requirements.txt
altgraph==0.17.2
astroid==2.12.5
boto3==1.24.64
botocore==1.27.64
certifi==2022.6.15
charset-normalizer==2.1.1
dill==0.3.5.1
idna==3.3
isort==5.10.1
jmespath==1.0.1
lazy-object-proxy==1.7.1
macholib==1.16
mccabe==0.7.0
mysql-connector-python==8.0.30
natura==0.0.65
numpy==1.23.2
pandas==1.4.4
platformdirs==2.5.2
protobuf==3.20.1
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.10
pylint==2.15.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
roman==3.3
s3transfer==0.6.0
six==1.16.0
slack-bolt==1.14.3
slack-sdk==3.18.1
tomli==2.0.1
tomlkit==0.11.4
urllib3==1.26.12
wrapt==1.14.1
